+++ pms-dashboard/vite.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
       '/v1': 'http://127.0.0.1:8080',
       '/admin': 'http://127.0.0.1:8080',
       '/l/': 'http://127.0.0.1:8080',
+      '/services': 'http://127.0.0.1:8443',
     }
   }
 })